refactor: moving filter and map to service
