+++ apps/namadillo/src/atoms/chain/atoms.ts
@@ -3,7 +3,6 @@ import namadaAssets from "@namada/chain-registry/namada/assetlist.json";
 import namada from "@namada/chains/chains/namada";
 import { IbcToken, NativeToken } from "@namada/indexer-client";
 import { indexerApiAtom } from "atoms/api";
-import { getDenomFromIbcTrace } from "atoms/integrations";
 import {
   defaultServerConfigAtom,
   indexerUrlAtom,
@@ -12,7 +11,6 @@ import {
 import { queryDependentFn } from "atoms/utils";
 import BigNumber from "bignumber.js";
 import * as osmosis from "chain-registry/mainnet/osmosis";
-import { findAssetByDenom } from "integrations/utils";
 import { atom } from "jotai";
 import { atomWithQuery } from "jotai-tanstack-query";
 import {
@@ -22,7 +20,7 @@ import {
   ChainStatus,
   MaspAssetRewards,
 } from "types";
-import { findAssetByToken, unknownAsset } from "utils/assets";
+import { findAssetByToken } from "utils/assets";
 import { calculateUnbondingPeriod } from "./functions";
 import {
   fetchChainParameters,
@@ -148,21 +146,7 @@ export const maspRewardsAtom = atomWithQuery((get) => {
   return {
     queryKey: ["masp-rewards", chain],
     ...queryDependentFn(async (): Promise<MaspAssetRewards[]> => {
-      const rewards = await fetchMaspRewards();
-      const existingRewards: MaspAssetRewards[] = rewards
-        .filter((r) => r.max_reward_rate > 0)
-        .map((r) => {
-          const denom = getDenomFromIbcTrace(r.name);
-          const asset = findAssetByDenom(denom) ?? unknownAsset(denom);
-          return {
-            asset,
-            kdGain: new BigNumber(r.kd_gain),
-            kpGain: new BigNumber(r.kp_gain),
-            lockedAmountTarget: new BigNumber(r.locked_amount_target),
-            maxRewardRate: new BigNumber(r.max_reward_rate),
-          };
-        });
-      return existingRewards;
+      return await fetchMaspRewards();
     }, [chain]),
   };
 });
